+++ verification-stubs/src/net/minecraft/util/Vec3.java
@@ -1,0 +1,5 @@
+package net.minecraft.util;
+public class Vec3 {
+    public final double xCoord,yCoord,zCoord;
+    public Vec3(double x,double y,double z){this.xCoord=x;this.yCoord=y;this.zCoord=z;}
+}